pyproject.toml
[project]
name = "darktable-ai"
version = "0.1.0"
description = "AI model conversion and packaging pipeline for darktable"
requires-python = ">=3.11,<3.13"
license = "GPL-3.0"
readme = "README.md"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.scripts]
dtai = "darktable_ai.cli:main"

[dependency-groups]
core = [
    "torch>=2.2.0,<2.7",
    "torchvision>=0.17.0,<0.22",
    "numpy<2.0",
    "onnx>=1.10.0",
    "onnxruntime",
    "onnxsim",
    "opencv-python",
]
nafnet = [
    {include-group = "core"},
    "onnxconverter-common",
    "addict",
    "lmdb",
    "scikit-image",
    "scipy",
    "tqdm",
    "yapf",
    "requests",
    "future",
    "tb-nightly",
]
nind = [
    {include-group = "core"},
    "onnxconverter-common",
]
rawnind = [
    {include-group = "core"},
    "onnxconverter-common",
    "rawpy",
    "tifffile",
    "configargparse",
    "pytorch-msssim",
    "colour-science",
    "ptflops",
]
segnext = [
    {include-group = "core"},
    "mmcv>=2.2.0",
    "albumentations",
    "easydict",
    "setuptools",
]
bsrgan = [
    {include-group = "core"},
]
sam21 = [
    {include-group = "core"},
    "torch>=2.5.1",
    "hydra-core>=1.3.2",
    "iopath>=0.1.10",
]
openclip = [
    {include-group = "core"},
    "open_clip_torch",
    "onnxscript",
]
all-models = [
    {include-group = "nafnet"},
    {include-group = "nind"},
    {include-group = "rawnind"},
    {include-group = "segnext"},
    {include-group = "bsrgan"},
    {include-group = "sam21"},
    {include-group = "openclip"},
]
eval = [
    "numpy<2.0",
    "onnxruntime",
    "Pillow",
    "scipy",
]

[tool.uv.extra-build-dependencies]
mmcv = ["setuptools<71"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["darktable_ai"]
